third_party/blink/web_tests/platform/generic/svg/custom/feDisplacementMap-invalid-selector-expected.txt

CONSOLE ERROR: Error: <feDisplacementMap> attribute yChannelSelector: Unexpected end of attribute. Unrecognized enumerated value, "".
